+++ tests/index.spec.js
@@ -1,0 +1,5 @@
+const { Game } = require("../src/app.js");
+
+test("two plus two is four", () => {
+  expect(2 + 2).toBe(4);
+});
